requirements.txt
dlib==19.24.0
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.2.0
pyserial==3.5
retry==0.9.2
pytest==7.2.0
flake8==5.0.4